Switch RoT builds to require RFD374 stage0.
- Removes the stage0/ program and its build configs, including CI

- Moves the DICE logic from stage0 into an `lpc55-rot-startup` library,
  which is in turn called by the RoT main function(s) before starting
  the kernel. This performs the same logic as the previous code,
  including reading a header out of the binary; we can improve this
  later by making more stuff happen at build time.

- Revs the header magic number, ensuring that the original stage0 will
  not attempt to boot one of these images (since it would potentially do
  so in nonsecure-mode, which wouldn't work). This also has the
  potential to allow in-place upgrades, though this is not currently
  being tested.
