+++ .github/workflows/dist.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
-        build: [stm32f3, stm32f4, lpc55, lpc55-stage0, stm32h743, stm32h753, gemini, rot-carrier, rot-carrier-stage0, gimletlet, gimlet-b, gimlet-b-lab, gimlet-c, gimlet-c-lab, gimlet-d, gimlet-d-lab, sidecar-b, sidecar-b-lab, sidecar-c, sidecar-c-lab, psc-b, psc-c, stm32g0, gimlet-rot-b, gimlet-rot-b-stage0, gimlet-rot-b-stage0-lab, gimlet-rot-c, gimlet-rot-c-stage0, gimlet-rot-c-stage0-lab, donglet-g031]
+        build: [stm32f3, stm32f4, lpc55, stm32h743, stm32h753, gemini, rot-carrier, gimletlet, gimlet-b, gimlet-b-lab, gimlet-c, gimlet-c-lab, gimlet-d, gimlet-d-lab, sidecar-b, sidecar-b-lab, sidecar-c, sidecar-c-lab, psc-b, psc-c, stm32g0, gimlet-rot-b, gimlet-rot-c, donglet-g031]
         include:
           - build: stm32g0
             app_name: demo-stm32g070-nucleo
@@ -33,11 +33,6 @@ jobs:
             app_toml: app/lpc55xpresso/app.toml
             target: thumbv8m.main-none-eabihf
             image: [a, b]
-          - build: lpc55-stage0
-            app_name: lpc55xpresso
-            app_toml: app/lpc55xpresso/stage0.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
           - build: stm32h743
             app_name: demo-stm32h743-nucleo
             app_toml: app/demo-stm32h7-nucleo/app-h743.toml
@@ -58,11 +53,6 @@ jobs:
             app_toml: app/rot-carrier/app.toml
             target: thumbv8m.main-none-eabihf
             image: [a, b]
-          - build: rot-carrier-stage0
-            app_name: rot-carrier
-            app_toml: app/rot-carrier/stage0.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
           - build: gimletlet
             app_name: gimletlet
             app_toml: app/gimletlet/app.toml
@@ -133,31 +123,11 @@ jobs:
             app_toml: app/gimlet-rot/app-b.toml
             target: thumbv8m.main-none-eabihf
             image: [a, b]
-          - build: gimlet-rot-b-stage0
-            app_name: gimlet-rot-b
-            app_toml: app/gimlet-rot/stage0-b.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
-          - build: gimlet-rot-b-stage0-lab
-            app_name: gimlet-rot-b-lab
-            app_toml: app/gimlet-rot/stage0-b-lab.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
           - build: gimlet-rot-c
             app_name: gimlet-rot-c
             app_toml: app/gimlet-rot/app-c.toml
             target: thumbv8m.main-none-eabihf
             image: [a, b]
-          - build: gimlet-rot-c-stage0
-            app_name: gimlet-rot-c
-            app_toml: app/gimlet-rot/stage0-c.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
-          - build: gimlet-rot-c-stage0-lab
-            app_name: gimlet-rot-c-lab
-            app_toml: app/gimlet-rot/stage0-c-lab.toml
-            target: thumbv8m.main-none-eabihf
-            image: stage0
           - build: donglet-g031
             app_name: donglet-g031
             app_toml: app/donglet/app-g031.toml
