+++ app/gimlet-rot/Cargo.toml
@@ -15,6 +15,8 @@ cfg-if = { workspace = true }
 
 abi = { path = "../../sys/abi"}
 kern = { path = "../../sys/kern" }
+lpc55-rot-startup = { path = "../../lib/lpc55-rot-startup" }
+unwrap-lite = { path = "../../lib/unwrap-lite" }
 
 # this lets you use `cargo fix`!
 [[bin]]
