--- app/gimlet-rot/stage0-b-lab.toml
@@ -1,34 +1,0 @@
-name = "gimlet-rot-b-lab"
-target = "thumbv8m.main-none-eabihf"
-board = "gimlet-rot-b"
-chip = "../../chips/lpc55"
-memory = "256k.toml"
-stacksize = 1024
-image-names = ["stage0"]
-external-images = ["a", "b"]
-
-[kernel]
-name = "stage0"
-requires = {flash = 0x9000, ram = 16000}
-features = ["dice-self"]
-stacksize = 13000
-
-[tasks.idle]
-name = "task-idle"
-priority = 0
-max-sizes = {flash = 256, ram = 256}
-stacksize = 256
-start = true
-
-[signing]
-enable-secure-boot = true
-enable-dice = true
-dice-inc-nxp-cfg = false
-dice-cust-cfg = false
-dice-inc-sec-epoch = false
-boot-error-gpio = { port = 0, pin = 17 } # ROT_TO_IGNIT_FLT_L
-
-[signing.certs]
-signing-certs = ["../../support/fake_certs/fake_certificate.der.crt"]
-root-certs = ["../../support/fake_certs/fake_certificate.der.crt"]
-private-key = "../../support/fake_certs/fake_private_key.pem"
