+++ app/lpc55xpresso/Cargo.toml
@@ -6,6 +6,7 @@ version = "0.1.0"
 
 [features]
 dump = ["kern/dump"]
+dice-self = ["lpc55-rot-startup/dice-self"]
 
 [dependencies]
 cortex-m = { workspace = true }
@@ -15,6 +16,8 @@ cfg-if = { workspace = true }
 
 abi = { path = "../../sys/abi"}
 kern = { path = "../../sys/kern" }
+lpc55-rot-startup = {path = "../../lib/lpc55-rot-startup"}
+unwrap-lite = {path = "../../lib/unwrap-lite"}
 
 # this lets you use `cargo fix`!
 [[bin]]
