+++ app/lpc55xpresso/src/main.rs
@@ -6,42 +6,17 @@
 #![no_main]
 
 use cortex_m_rt::entry;
-
-// When we're secure we don't have access to read the CMPA/NMPA where the
-// official setting is stored, emulate what the clock driver does instead
-fn get_clock_speed() -> (u32, u8) {
-    // We need to set the clock speed for flash programming to work
-    // properly. Reading it out of syscon is less error prone than
-    // trying to compile it in at build time
-
-    let syscon = unsafe { &*lpc55_pac::SYSCON::ptr() };
-
-    let a = syscon.mainclksela.read().bits();
-    let b = syscon.mainclkselb.read().bits();
-    let div = syscon.ahbclkdiv.read().bits();
-
-    // corresponds to FRO 96 MHz, see 4.5.34 in user manual
-    const EXPECTED_MAINCLKSELA: u32 = 3;
-    // corresponds to Main Clock A, see 4.5.45 in user manual
-    const EXPECTED_MAINCLKSELB: u32 = 0;
-
-    // We expect the 96MHz clock to be used based on the ROM.
-    // If it's not there are probably more (bad) surprises coming
-    // and panicking is reasonable
-    if a != EXPECTED_MAINCLKSELA || b != EXPECTED_MAINCLKSELB {
-        panic!();
-    }
-
-    if div == 0 {
-        (96, div as u8)
-    } else {
-        (48, div as u8)
-    }
-}
+use lpc55_rot_startup::{get_clock_speed, startup};
+use unwrap_lite::UnwrapLite;
 
 #[entry]
 fn main() -> ! {
-    let (cycles_per_ms, _div) = get_clock_speed();
+    let core_peripherals = cortex_m::Peripherals::take().unwrap_lite();
+    let peripherals = lpc55_pac::Peripherals::take().unwrap_lite();
+
+    let (cycles_per_ms, div) = get_clock_speed(&peripherals);
+
+    startup(&core_peripherals, &peripherals);
 
     unsafe { kern::startup::start_kernel(cycles_per_ms * 1_000) }
 }
