--- app/lpc55xpresso/stage0.toml
@@ -1,32 +1,0 @@
-name = "lpc55xpresso"
-target = "thumbv8m.main-none-eabihf"
-board = "lpcxpresso55s69"
-chip = "../../chips/lpc55"
-image-names = ["stage0"]
-external-images = ["a", "b"]
-
-[kernel]
-name = "stage0"
-requires = {flash = 0x9000, ram = 16000}
-features = ["dice-self"]
-stacksize = 13000
-
-[tasks.idle]
-name = "task-idle"
-priority = 0
-max-sizes = {flash = 256, ram = 256}
-stacksize = 256
-start = true
-
-[signing]
-enable-secure-boot = true
-enable-dice = true
-dice-inc-nxp-cfg = false
-dice-cust-cfg = false
-dice-inc-sec-epoch = false
-boot-error-gpio = { port = 1, pin = 1 } # BOOT0_LED
-
-[signing.certs]
-signing-certs = ["../../support/fake_certs/fake_certificate.der.crt"]
-root-certs = ["../../support/fake_certs/fake_certificate.der.crt"]
-private-key = "../../support/fake_certs/fake_private_key.pem"
