+++ lib/lpc55-rot-startup/build.rs
@@ -2,28 +2,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use std::fs::File;
-use std::io::Write;
-
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let out = build_util::out_dir();
-    let mut const_file = File::create(out.join("consts.rs")).unwrap();
-
-    let image_id: u64 = build_util::env_var("HUBRIS_IMAGE_ID")?.parse()?;
-
-    writeln!(const_file, "// See build.rs for details")?;
-
-    writeln!(const_file, "#[used]")?;
-    writeln!(const_file, "#[no_mangle]")?;
-    writeln!(const_file, "#[link_section = \".hubris_id\"]")?;
-    writeln!(
-        const_file,
-        "pub static HUBRIS_IMAGE_ID: u64 = {};",
-        image_id
-    )?;
-
     #[cfg(feature = "dice-mfg")]
-    gen_memory_range(&out)?;
+    {
+        let out = build_util::out_dir();
+        gen_memory_range(&out)?;
+    }
 
     Ok(())
 }
@@ -33,6 +17,7 @@ fn gen_memory_range(
     out_dir: &std::path::PathBuf,
 ) -> Result<(), Box<dyn std::error::Error>> {
     use serde::Deserialize;
+    use std::fs::File;
 
     #[derive(Deserialize, Debug)]
     struct DiceMfgRegion {
