+++ lib/lpc55-rot-startup/src/handoff.rs
